+++ config/routes.rb
@@ -1,0 +1,2 @@
+get 'qualification', :to => 'qualification#index'
+post 'qualification', :to => 'qualification#enable'